Fix error copied from the source paper
The models were based on Populus manual that had an error in
equation dN/dt/N in one place. It could have been possible to
find this error by careful reading, since the term was correct
elsewhere. I checked the model against Tilman 1982 to find the
source of the problem. Now the results are similar to the Populus
manual instead of having halved population sizes.